+++ plugins/nunjucks.js
@@ -5,5 +5,10 @@ export default function () {
     const nunjuksEngine = new NunjuksEngine(site);
 
     site.engine([".njk", ".html"], nunjuksEngine);
+    site.filter("njk", filter);
+
+    function filter(string, data = {}) {
+      return nunjuksEngine.engine.renderString(string, data);
+    }
   };
 }